boards/samr21-xpro: move EUI provider config to eui_provider_params.h
